hw15 Трансформаторы

+++ hw15-spring-integration/src/main/java/ru/otus/hw/domain/Pupa.java
@@ -1,3 +1,3 @@
 package ru.otus.hw.domain;
 
-public record Pupa(String name, long weight) { }
+public record Pupa(String name, double weight) { }
